changes on places assuming int=sint32 (while BeOS headers sint32=long)
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1262 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/components/gui_speedbar.h
@@ -34,7 +34,7 @@ private:
 public:
 	gui_speedbar_t() { base = 100; vertical = false; }
 
-	void add_color_value(const sint32 *value, int color);
+	void add_color_value(const sint32 *value, uint8 color);
 
 	void set_base(sint32 base);
